client: add Enclave type
This commit adds the `Enclave` type.
An Enclave is an isolated area within
a KES server and implements multi-tenancy.

This commit adds a minimal client-side
implementation of enclaves.

Currently, the KES server does not support
multi-tenancy. However, future versions
(stateful servers) will support multiple
enclaves.
